Setting up configs

+++ CustomItems/Items/LethalInjection.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using CustomItems.API;
 using Exiled.API.Features;
 using Exiled.Events.EventArgs;
@@ -12,7 +13,8 @@ namespace CustomItems.Items
         }
 
         public override string ItemName { get; set; } = "LI-119";
-
+        public override Dictionary<SpawnLocation, float> SpawnLocations { get; set; } =
+            Plugin.Singleton.Config.ItemConfigs.LethalCfg.SpawnLocations;
         protected override string ItemDescription { get; set; } =
             "This is a Lethal Injection that, when used, will cause SCP-096 to immediately leave his enrage, regardless of how many targets he currently has, if you are one of his current targets. You always die when using this, even if there's no enrage to break, or you are not a target.";
 
@@ -35,12 +37,20 @@ namespace CustomItems.Items
                     {
                         if (player.CurrentScp is PlayableScps.Scp096 scp096)
                         {
-                            if (scp096.HasTarget(ev.Player.ReferenceHub) && scp096.PlayerState == Scp096PlayerState.Enraged || scp096.PlayerState == Scp096PlayerState.Enraging|| scp096.PlayerState == Scp096PlayerState.Attacking)
+                            if (scp096.HasTarget(ev.Player.ReferenceHub) &&
+                                scp096.PlayerState == Scp096PlayerState.Enraged ||
+                                scp096.PlayerState == Scp096PlayerState.Enraging ||
+                                scp096.PlayerState == Scp096PlayerState.Attacking)
+                            {
                                 scp096.ResetEnrage();
+                                ev.Player.Kill(DamageTypes.Poison);
+                                return;
+                            }
                         }
                     }
                 
-                ev.Player.Kill(DamageTypes.Poison);
+                if (Plugin.Singleton.Config.ItemConfigs.LethalCfg.KillOnFail)
+                    ev.Player.Kill(DamageTypes.Poison);
             }
         }
     }
